Test: RacingCarGame 테스트 추가
