+++ src/test/java/racingcar/ApplicationTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test;
 import static camp.nextstep.edu.missionutils.test.Assertions.assertRandomNumberInRangeTest;
 import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
 class ApplicationTest extends NsTest {
     private static final int MOVING_FORWARD = 4;
@@ -24,24 +24,10 @@ class ApplicationTest extends NsTest {
     }
 
     @Test
-    void 게임_3명_5회() {
-        assertRandomNumberInRangeTest(
-                ()->{
-                    run("pobi,woni,jun", "5");
-                    assertThat(output()).contains(
-                            "pobi : -", "woni : ","jun : -"
-                            ,"pobi : --","woni : -","jun : --"
-                            ,"pobi : ---","woni : --","jun : ---"
-                            ,"pobi : ----","woni : ---","jun : ----"
-                            ,"pobi : -----","woni : ----","jun : -----"
-                            ,"최종 우승자 : pobi, jun"
-                    );
-                },
-                MOVING_FORWARD, STOP, MOVING_FORWARD
-                ,MOVING_FORWARD, MOVING_FORWARD, MOVING_FORWARD
-                ,MOVING_FORWARD, MOVING_FORWARD, MOVING_FORWARD
-                ,MOVING_FORWARD, MOVING_FORWARD, MOVING_FORWARD
-                ,MOVING_FORWARD, MOVING_FORWARD, MOVING_FORWARD
+    void 이름에_대한_예외_처리() {
+        assertSimpleTest(() ->
+                assertThatThrownBy(() -> runException("pobi,javaji", "1"))
+                        .isInstanceOf(IllegalArgumentException.class)
         );
     }
